feat: make it possible to define external hostnames for nodes
This is possible by defing a new entry in the replsets section in the crd. Here is an example for a
cluster named "smurf" with replicaset name rs0:

    Connectivity:
      Mapping:
        External:  n0.smurf.mongodb.example.com
        Internal:  smurf-rs0-0
        External:  n1.smurf.mongodb.example.com
        Internal:  smurf-rs0-1
        External:  n2.smurf.mongodb.example.com
        Internal:  smurf-rs0-2

This will cause the following things to happen:

- SANs of generated certificates will also contain those external hostnames
- Each service will get an annotated with `psmdbExternalHostName: n0.smurf.mongodb.example.com` (depending on external hostname)
- Hostnames of the mongodb nodes will be set to the external hostname

To make this setup fully work, a DNS horizon has to be created inside the cluster, so the nodes inside the cluster connect directly to each other.
This might be for example done by configuring CoreDNS like this (it also generalizes the namespace)

    rewrite stop {
      name regex n(.*)\.(.*)\.(.*)\.example.com\.$ {2}-rs0-{1}.{3}.svc.cluster.local
      answer name (.*)-rs0-(.*)\.(.*)\.svc\.cluster\.local\.$ n{2}.{1}.{3}.example.com
    }

After this you can route tcp traffic from the outside to the cluster using an ingress controller. Here is a part of traefik ingress configuration:

    Routes:
      Match:  HostSNI(`n0.smurf.mongodb.example.com`)
      Services:
        Name: smurf-rs0-0
        Port: 27017

… and for the other nodes as well. This enabled clients to connect to each node and get full connectivity with sharing load balancers between multiple clusters.