K8SPSMDB-455: Fix rollback after major upgrade
If cluster upgraded to next major version with setFCV false, mongod
binaries replaced with new version but FeatureCompabilityVersion remains
unchanged. If user tries to rollback to previous version operator won't
allow the operation due to our checks and user is forced to manually
set FCV twice. This change allows user to update the cluster to same
version as FCV.