K8SPSMDB-506: Fix restarting primary pod on smart update (#719)
* K8SPSMDB-506: Fix restarting primary pod on smart update

We try to decide which pod is primary by running `rs.status()` and
getting `name` field of PRIMARY. If replset is not exposed `name` is
local FQDN of service and since service name and pod name is the same we
can select the right pod. But if replset is exposed node names are
external hostnames and selecting the primary pod is not straightforward.
We need to get the external address of each pod in the replset and
compare with primary host.