K8SPSMDB-265: Configure backup node priorities for PBM
PBM 1.5.0 introduced a new feature to influence the backup node
selection. It selects the node with highest priority in the config.

As part of cross cluster replication changes, we need to make sure we
start backups on nodes that are part of the operator's own cluster. With
this commit, we're assigning lower priority (0.5) to external nodes so
they won't be selected by PBM.

According to PBM's implementation, if you pass a subset of the nodes,
all remaining nodes will be automatically considered as priority 1
including the primary. It's not ideal, since we don't want to take
backups on primary (which why the PBM itself assigns priority 0.5 to the
primary node if user doesn't specify anything) and we don't want to
change the backup behavior for the users. That's also the reason of the
changes about mongo client functions. It was only possible to use them
from the main controller, but I needed to initialize a client in both
backup and restore controllers.