K8SPSMDB-640: Fix updating custom labels
We were passing custom labels defined in CR to
PersistentVolumeClaimTemplate in StatefulSets to fix K8SPSMDB-437.
Although it's working for the *initial deploy*, if the user updates the
custom labels the reconciliation stuck because it's not allowed to
update PVC template in StatefulSet.

To fix this problem, this commit introduces a new function that patches
PVCs to update their labels in every reconciliation and ensure
StatefulSet labels match the PVC labels.