K8SPSMDB-656 fix: missing cluster name in backup `additionalPrinterColumns`
https://jira.percona.com/browse/K8SPSMDB-656

There are two fields for a cluster name in backup. `psmdbCluster` is a deprecated and `clusterName` is a new one. There is no way to display the first field if second one is empty in `additionalPrinterColumns`. That's why we should replace `clusterName` with `psmdbCluster` if the first one is empty after creation