K8SPSMDB-533: special chars for users breaks the cluster
https://jira.percona.com/browse/K8SPSMDB-533

Passwords were stored url encoded in the secrets. Because of that, we create users with passwords that don't contain special chars, but url encoded. That would cause the problem if we try to connect to mongo without using the connection string. We should store passwords not encoded.

Also, we use shell when creating `userAdmin` user in `handleReplsetInit` function. If the password contains shell special chars or quotes, that would break the creation of the user.

Finally, when using cr version < 1.2.0, we use the mongo connection string and try to get credentials from secrets. That's why I added url encoded username and password to the secret with `_ESCAPED` suffix