K8SPSMDB-651: Recover from full cluster crash (#902)
* K8SPSMDB-651: Recover from full cluster crash

If all pods of replica set removed when the cluster is exposed with
LoadBalancer, they fail the join to the replica set again.

If replset is exposed, the operator configures the replset to use
external IP of each load balancer. If all pods are deleted together,
they can't understand if they're part of the replset or not because they
know nothing about load balancer (we're using --bind_ip_all). The
operator sees the replica set as initialized but fail to connect to it
due to ReplicaSetNoPrimary.

The fix:

Rather than removing members to reach majority (members we deleted may
have more up to date data) or increasing the priority of a random member
until it gets elected (some other member may have more up to date data),
the operator will reconfigure the replset to use local FQDNs and it'll
recover the cluster from the crash. We'll only perform recovery in case
of server selection error with topology ReplsetNoPrimary and if replset
is exposed and in the next reconciliation loops the operator reconfigure
the members again to use external IPs.

* pass context

* fix arbiter test

Co-authored-by: Viacheslav Sarzhan <slava.sarzhan@percona.com>