K8SPSMDB-724: fix `delete-backup` finalizer
https://jira.percona.com/browse/K8SPSMDB-724

Backup deletion freezes if cr is deleted before the backup. To delete backups, we used PBM's `DeleteBackup` method, which deletes backup records and files. But if there is no cr, we should only delete files. Using `DeleteBackup` is not an option in this situation because it requires a connection to the database. Instead, we should use the `DeleteBackupFiles` method.