K8SPSMDB-730: Fix enabling PITR unnecessarily (#1002)
* K8SPSMDB-730: Fix enabling PITR unnecessarily

3eb140a41c58dd6f2bce56c963afc4ef4e2462d4 introduced a fix to properly
set PITR compression variables in PBM config. PBM needs disabling and
enabling the PITR for new values to take effect, the commit implements
this but doesn't consider if PITR is enabled or not in the CR itself.
This makes the operator enable PITR momentarily (until the next
reconciliation where operator detect it's disabled in CR and disable it
in PBM config too). PBM creates a lock for PITR in `pbmLock` collection
but after the operator disables PITR, PBM doesn't release this lock and
it blocks restores in the cluster. This commit ensures we only set
compression variables if PITR enabled.

* fix databaseAdmin on upgrade

* fix databaseAdmin on upgrade

* retry killing containers on error

* use GKE 1.21 on Jenkins

Co-authored-by: Viacheslav Sarzhan <slava.sarzhan@percona.com>