K8SPSMDB-666: Don't force reconfig by default (#980)
* K8SPSMDB-666: Don't force reconfig by default

We introduced `force` option to `replSetReconfig` with PSMDB v4.4 support
to conform the rule:

> A new configuration can make **at most one** of the following changes to
> the cluster membership:
>
> * Adding a new voting replica set member.
> * Removing an existing voting replica set member.
> * Modifying the votes for an existing replica set member.

This commit removes the `force` option and ensures we do only one of the
permitted changes above in every reconfiguration. For example, if the
user scale the ReplicaSet from 3 to 5, the operator is going to add new
pods one by one to the ReplicaSet configuration.

Removing `force` also forced us to remove old nodes and fix
priority/votes in two separate `replSetReconfig` operations. If the user
scale down the ReplicaSet from 5 to 3, the operator will delete the last
pod (i.e cluster1-rs0-4) and then strip the voting rights from the other
(i.e cluster1-rs0-3) since there's a even number of voters at the
moment. This is not permitted since it breaks the rule above. This means
there'll be a time window that there are even number of voters in the
ReplicaSet config, although duration of this window is really short.

* fix adding arbiters to replset config

Co-authored-by: Viacheslav Sarzhan <slava.sarzhan@percona.com>