K8SPSMDB-796: Fix PBM connection if replset is exposed (#1060)
If replset is exposed, the operator uses ClusterIPs in the replset
config. When PBM tries to open connection to replset using TLS
certificates, it fails since these IPs are not included in the
certificate. Ideally we want to move away from using unsafe options from
TLS but it needs careful thinking.