K8SPSMDB-867: Use local FQDNs unless DNS mode is external
We were using exposed IPs in replset configuration in case replset is
exposed. This was causing occasional full cluster crashes, hard to
reason about and apparently against the MongoDB best practices.

Having exposed IP addresses in replset configuration is especially
important for enabling multi cluster deployments. In a multi cluster
setup, replset members can be in different Kubernetes cluster but they
need to be able to reach each other to for replication to work. At the
beginning, exposing members to the internet and configuring the replset
with exposed IPs was the only solution to ensure connectivity between
members. However the community developed their own ways to solve this
problem, namely using service meshes. We also integrated the operator
with Multi-Cluster Service APIs to enable multi cluster deployments
without exposing the database to the internet. Now we can say using
exposed IP addresses is not the safest way to have multi cluster
topologies.

With these changes, we're going to use local FQDNs in replset
configuration even replset is exposed by default. Users will be able to
configure this behavior using `clusterServiceDNSMode` field. We're
adding a new value to this field: `External` but for the sake of
completeness, let's reiterate all possible values:

1. **`Internal`**: Use local FQDNs (i.e.
   `cluster1-rs0-0.cluster1-rs0.psmdb.svc.cluster.local`) in replset
   configuration even if the replset is exposed. **This is the default
   value.**
2. **`ServiceMesh`**:  Use a special FQDN using the pod name (i.e.
   `cluster1-rs0-0.psmdb.svc.cluster.local`) assuming it's resolvable
   and available in all clusters.
3. **`External`**:  Use exposed IP in replset configuration if replset
   is exposed. Otherwise use local FQDN. **This is basically the same
   with the previous behavior.**