K8SPSMDB-733: Improve failure detection for PVC resize
Operator were reporting that PVC resize completed successfully even
though it's failed. This commit fixes this behavior and introduce some
improvements.

Operator tries to understand if resize operation in progress by checking
the status.conditions of PVCs. If the condition says it's resizing
the operation is in progress, if there's no condition then the operation
completed successfully. This approach is naive since resize could be
failed. With these changes operator will check events regarding to
the PVC and check if `VolumeResizeFailed` event exists. If it does,
resize operation will declared as failed.

If a resize operation fails, operator will automatically revert the PVC
size in cr.yaml. This way the probability to have discrepancy between `CR <->
STS <-> PVC` will be lower. If resize operation fails only for some
PVCs, the operator will still revert the PVC size in cr.yaml with the
size of failed PVCs.