K8SPSMDB-1088: Prevent duplicate backup commands
This is an attempt to fix an edge case: If user patches the backup
object after operator sent backup command to PBM but object's status is
still New, operator might send another backup command to PBM. These
changes can't guarantee that this issue 100% fixed but they
significantly lower the probability of hitting this bug.

Also this commit fixes the unnecessarily requeued reconciles for
finished backups.